+++ concordium-std/src/test_infrastructure.rs
@@ -908,7 +908,6 @@ impl HasCryptoPrimitives for TestCryptoPrimitives {
     ) -> bool {
         #[cfg(feature = "crypto-primitives")]
         {
-            use std::convert::TryFrom;
             let signature = ed25519_zebra::Signature::try_from(&signature.0[..]);
             let public_key = ed25519_zebra::VerificationKey::try_from(&public_key.0[..]);
             match (signature, public_key) {
@@ -1289,7 +1288,7 @@ pub struct TestHost<State> {
     /// Functions that mock responses to calls.
     // This is Rc+RefCell because it needs to be cloneable. There might be another way to make the
     // MockFn cloneable, but this seemed like the easiest option.
-    mocking_fns:             Rc<RefCell<MockFnMap<State>>>,
+    mocking_fns: Rc<RefCell<MockFnMap<State>>>,
     /// Transfers the contract has made during its execution.
     transfers:               RefCell<Vec<(AccountAddress, Amount)>>,
     /// The contract balance. This is updated during execution based on contract
@@ -2407,9 +2406,9 @@ mod test {
         outer_map.insert(0u8, inner_map_1);
         outer_map.insert(1u8, inner_map_2);
         outer_map.clear();
-        let iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
+        let mut iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
         // The only remaining node should be the state_builder's next_item_prefix node.
-        assert!(iter.skip(1).next().is_none());
+        assert!(iter.nth(1).is_none());
     }
 
     #[test]